backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
PyJWT==2.9.0
ffmpeg-python==0.2.0
